requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-aiplatform
google-adk
python-dotenv
